refact: moved relevant methods to UserResponseUtil
